(index) Route position bitmask from converter to index
This commit changes the term metadata column from a compressed byte array to a long array.  Each long consists of a byte containing the old term metadata, and then a 56 bit mask containing approximate position data.

At this point, the index constructor just extracts the term metadata byte and proceeds with the original code.

+++ code/index/test/nu/marginalia/index/IndexQueryServiceIntegrationSmokeTest.java
@@ -413,7 +413,7 @@ public class IndexQueryServiceIntegrationSmokeTest {
         ));
 
         List<String> keywords = IntStream.of(factors).mapToObj(Integer::toString).toList();
-        byte[] metadata = new byte[factors.length];
+        long[] metadata = new long[factors.length];
         for (int i = 0; i < factors.length; i++) {
             metadata[i] = WordFlags.Title.asBit();
         }
@@ -453,7 +453,7 @@ public class IndexQueryServiceIntegrationSmokeTest {
 
 
         List<String> keywords = IntStream.of(factors).mapToObj(Integer::toString).toList();
-        byte[] metadata = new byte[factors.length];
+        long[] metadata = new long[factors.length];
         for (int i = 0; i < factors.length; i++) {
             metadata[i] = WordFlags.Title.asBit();
         }
